+++ src/main/java/org/montclairrobotics/cyborg/devices/CBIEncoder.java
@@ -1,11 +1,8 @@
 package org.montclairrobotics.cyborg.devices;
 
-import edu.wpi.first.wpilibj.*;
-import edu.wpi.first.wpilibj.smartdashboard.SendableBuilder;
-
 public interface CBIEncoder {
 
-    int getRaw();
+    double getRaw();
 
     void reset();
